requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
django-heroku==0.3.1
future==0.18.2
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
whitenoise==5.2.0
certifi==2020.11.8
chardet==3.0.4
